+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Try to reconnect to the database if necessary when cross-matching with Gaia.
 * Add `--fit-all-detections` flag for `cherno acquire` (default to `acquisition.fit_all_detections`) to adjust whether all detections or only the centre of each camera are used.
+* `cherno get-scale` now calculates the weighted average, giving higher weigth to recent scale measurements.
 
 ### 🏷️ Changed